lab9_airflow_part_2/21130791_NguyenThiNga_lab09/requirements.txt
# Airflow
apache-airflow==2.7.0

# Database Drivers
psycopg2-binary==2.9.3  # PostgreSQL connector
pymongo==4.1.1  # MongoDB connector

# Other packages
pandas==1.3.3